chat, revision
