--- Sockets-Exp-master/ChatApp/README.md
@@ -1,15 +1,0 @@
-# A Chat App
-	Aim: To build an app that demonstrates use of sockets to build a basic chat app to have realtime conversation with people online.
-
-## Getting Started
-
-- Clone repository using the link [https://github.com/Worm4047/Sockets-Exp.git](https://github.com/Worm4047/Sockets-Exp.git).
-- Install Flask-SocketIO using command `pip install flask-socketio`
-- Run `python app.py`
-- Open browser and navigate to localhost:5000
-- Off you go !!!
-
-
-
-
-
