--- Sockets-Exp-master/README.md
@@ -1,19 +1,0 @@
-# SyncExp
-Aim: To build web apps powered by Sockets.
-
-## Requirements
-
-1. *Flask-SocketIO*
-2. *Jquery + Socket.io*
-3. *Free time and willingness to learn*
-
-## What are we planning to build.
-
-- [x] A synchronised slider (Synchronized across multiple clients).
-- [ ]  A simple chat app to chat with people what are currently online.
-- [ ] A synchronised video player.
-- [ ] A chat room.
-
-
-So buckle up, it's going to be a fun ride.
-
