--- Sockets-Exp-master/Synchronized Slider/README.md
@@ -1,14 +1,0 @@
-# Synchronized Slider
-	Aim: To build an app that demonstrates use of sockets to send data back and forth between clients and server.
-
-## Getting Started
-
-- Clone repository using the link [https://github.com/Worm4047/Sockets-Exp.git](https://github.com/Worm4047/Sockets-Exp.git).
-- Install Flask-SocketIO using command `pip install flask-socketio`
-- Run `python app.py`
-- Open browser and navigate to localhost:5000
-
-
-
-
-
