+++ app/templates/chat.html
@@ -52,7 +52,7 @@
         </div></div>
         <ul class="list-unstyled messages-container">
           {% if team is defined %}
-          {% for message in team.messages.limit(10) %}
+          {% for message in team.last_messages(20)[::-1] %}
           <li class="d-flex justify-content-between mb-4">
             <img src="{{url_for('static', filename='images/profile.png')}}" alt="avatar"
               class="rounded-circle d-flex align-self-start me-3 shadow-1-strong" width="60">
@@ -73,12 +73,14 @@
           {% endif %}
         </ul>
         <ul class="list-unstyled">
+          <form id="send-message" method="post" action="#">
           <li class="bg-white mb-3">
             <div class="form-outline">
               <textarea class="form-control" id="message-text" rows="4"></textarea>
             </div>
           </li>
-          <button type="button" class="btn btn-info btn-rounded float-end">Send</button>
+          <input type="submit" class="btn btn-info btn-rounded float-end" value="Enviar">
+        </form>
         </ul>
       </div>
       {% endif %}
@@ -90,21 +92,24 @@
 
 {% block scripts %}
 {{ super() }}
-<script type="text/javascript" src="//code.jquery.com/jquery-1.4.2.min.js"></script>
-<!-- <script src="{{url_for('static', filename='js/chat.js')}}"></script> -->
+<script src="https://cdnjs.cloudflare.com/ajax/libs/jquery/3.5.1/jquery.min.js" integrity="sha512-bLT0Qm9VnAYZDflyKcBaQ2gg0hSYNQrJ8RilYldYQ1FxQYoCLtUjuuRuZo+fjqhx/qtq/1itJ0C2ejDxltZVFg==" crossorigin="anonymous"></script>
+
 <!-- <script type="text/javascript" src="//cdn.socket.io/4.4.1/socket.io.min.js"></script> -->
 <script src="https://cdnjs.cloudflare.com/ajax/libs/socket.io/3.0.4/socket.io.js" integrity="sha512-aMGMvNYu8Ue4G+fHa359jcPb1u+ytAF+P2SCb+PxrjCdO3n3ZTxJ30zuH39rimUggmTwmh2u7wvQsDTHESnmfQ==" crossorigin="anonymous"></script>
-<script type="text/javascript" charset="utf-8">
+<script src="{{url_for('static', filename='js/chat.js')}}"></script>
+
+
+{#<script type="text/javascript" charset="utf-8">
   var socket;
   $(document).ready(function(){
       //socket = io.connect('http://' + document.domain + ':' + location.port + '/chat');
-      socket = io();
+      var socket = io();
       socket.on('connect', function() {
         socket.emit('my_event', {data: 'I\'m connected!'});
+        console.log("Conectado")
     });
 
     socket.on('my_response', function(msg, cb) {
-      console.log(msg)
       $('#log').append('<br>' + $('<div/>').text('Received #' + msg.count + ': ' + msg.data).html());
       if (cb)
           cb();
@@ -174,5 +179,5 @@
       });
   }
 </script>
-
+#}
 {% endblock %}
